requirements.txt
asgiref==3.7.2
astroid==2.15.6
click==8.1.6
colorama==0.4.6
dill==0.3.7
Django==4.2.4
django-cors-headers==4.2.0
djangorestframework==3.14.0
docstring-generator==0.2.0
iniconfig==2.0.0
isort==5.12.0
lazy-object-proxy==1.9.0
mccabe==0.7.0
packaging==23.1
platformdirs==3.10.0
pluggy==1.2.0
pylint==2.17.5
pytest==7.4.0
pytest-django==4.5.2
pytz==2023.3
sqlparse==0.4.4
strongtyping==3.11.4
tomlkit==0.12.1
tzdata==2023.3
wrapt==1.15.0
